pyproject.toml
[tool.poetry]
name = "library-system-api"
version = "0.1.0"
description = "Library System API"
authors = ["Rodrigo Saraiva <rodrigobezerrasaraiva@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
djangorestframework = "^3.14.0"
pillow = "^9.4.0"
psycopg2-binary = "^2.9.5"


[tool.poetry.group.dev.dependencies]
isort = "^5.11.4"
mkdocs = "^1.4.2"
blue = "^0.9.1"
django-extensions = "^3.2.1"

[tool.isort]
profile = "black"
line_length = 79

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
